Refactor Netflix scraper package and introduce LeavingShowScraper
- Removed the old NetflixScraper initialization and replaced it with a new service for scraping leaving titles.
- Updated main.py to import StreamingAvailabilityScraper from the new leaving_show_scraper module.
- Added a new service module (leaving_show_scraper.py) that implements functionality to scrape and save leaving titles from the Streaming Availability API.
- Introduced logging and error handling in the new scraper for better debugging and data management.
