--- scraper/__init__.py
@@ -1,6 +1,0 @@
-"""
-Netflix scraper package initialization
-"""
-from .netflix_scraper import NetflixScraper
-
-__all__ = ['NetflixScraper']
